refactor team columns folder
