+++ shared-ui/components/meet-the-team/MeetTheTeamSection.tsx
@@ -1,32 +1,18 @@
 import React from 'react';
 import { teamInfo } from '../../lib/data';
-import { min } from '../../lib/responsive';
 import { TeamColumnInfo } from '../../lib/types';
 import { H2 } from '../../style/typography';
-import useMatchMedia from 'react-use-match-media';
-import {
-  StyledMeetTheTeamSection,
-  StyledImageContainer
-} from './MeetTheTeamSection.styles';
-import MobileTeamColumn from './mobile-team-column/MobileTeamColumn';
+import { StyledMeetTheTeamSection } from './MeetTheTeamSection.styles';
 import TeamColumn from './team-column/TeamColumn';
 import MeetAllCore from './meet-all-core/MeetAllCore';
 
 const MeetTheTeamSection: React.FC = () => {
   const teamColumnsInfo: TeamColumnInfo[] = teamInfo;
-  const isDesktop = useMatchMedia(min.tablet);
   return (
     <div id="team">
       <StyledMeetTheTeamSection>
         <H2>Meet the Team</H2>
-        {isDesktop && (
-          <StyledImageContainer>
-            {teamColumnsInfo.map((teamColumn: TeamColumnInfo) => (
-              <TeamColumn columnInfo={teamColumn} />
-            ))}
-          </StyledImageContainer>
-        )}
-        {!isDesktop && <MobileTeamColumn listOfColumnInfo={teamColumnsInfo} />}
+        <TeamColumn listOfColumnInfo={teamColumnsInfo} />
         <MeetAllCore />
       </StyledMeetTheTeamSection>
     </div>
