+++ shared-ui/components/meet-the-team/team-column/mobile-team-column/MobileTeamColumn.tsx
@@ -1,7 +1,7 @@
 import React, { useState } from 'react';
-import Arrow from '../../arrow/Arrow';
-import { MobileTeamColumnProps, TeamColumnInfo } from '../../../lib/types';
-import { getLeftOrRight } from '../../../../main-site/lib/utils';
+import Arrow from '../../../arrow/Arrow';
+import { MobileTeamColumnProps, TeamColumnInfo } from '../../../../lib/types';
+import { getLeftOrRight } from '../../../../../main-site/lib/utils';
 import {
   StyledMobileTeamColumn,
   LabelArrowContainer,
